refactor(tmux/mouse): move webview pointer routing into webview.rs
